requirements.txt
annotated-types==0.7.0
asyncio==3.4.3
build==1.2.1
colorama==0.4.6
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
pyproject_hooks==1.1.0
PyYAML==6.0.1
ruff==0.5.0
typing_extensions==4.12.2
